LeaderBoard Times Fixes
Leaderboard Time laat zich nu goed zien allen nog max 1 entry morgen met andere vr testen + build

+++ Assets/Floris/Scripts/LeaderBoard/LeaderboardCreator/Scripts/Main/LeaderboardCreator.cs
@@ -128,7 +128,7 @@ namespace Dan.Main
         /// <param name="score">The highscore of the player</param>
         /// <param name="callback">Returns true if the request was successful.</param>
         /// <param name="errorCallback">Returns an error message if the request failed.</param>
-        public static void UploadNewEntry(string publicKey, string username, int score, Action<bool> callback = null, Action<string> errorCallback = null) => 
+        public static void UploadNewEntry(string publicKey, string username, float score, Action<bool> callback = null, Action<string> errorCallback = null) => 
             UploadNewEntry(publicKey, username, score, " ", callback, errorCallback);
 
         /// <summary>
@@ -140,7 +140,7 @@ namespace Dan.Main
         /// <param name="extra">Extra data to be stored with the entry (max length of 100, unless using an advanced leaderboard)</param>
         /// <param name="callback">Returns true if the request was successful.</param>
         /// <param name="errorCallback">Returns an error message if the request failed.</param>
-        public static void UploadNewEntry(string publicKey, string username, int score, string extra, Action<bool> callback = null, Action<string> errorCallback = null)
+        public static void UploadNewEntry(string publicKey, string username, float score, string extra, Action<bool> callback = null, Action<string> errorCallback = null)
         {
             if (string.IsNullOrEmpty(publicKey))
             {
